requirements.txt
pillow
requests
pypdf